refactor(mobile): split share sheet and add analytics test seam
- AnalyticsService: the singleton behind the factory is now swappable via
  a visibleForTesting setter, giving every existing AnalyticsService()
  call site a test seam without changing the call convention.
- Extract the forum filter tab (selfie acquisition, consent, wall
  submission) into clean_air_forum_filter_tab.dart; selfie and consent
  state stay lifted in the sheet so they survive tab switches.
- Move shared presentational widgets and the RepaintBoundary capture
  helper into share_sheet_widgets.dart.

No behavior changes; air_quality_share_sheet.dart drops from 916 to 420
lines.
