+++ src/mobile/lib/src/app/dashboard/widgets/air_quality_share_sheet.dart
@@ -1,24 +1,19 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:typed_data';
-import 'dart:ui' as ui;
 
 import 'package:airqo/src/app/dashboard/models/airquality_response.dart';
 import 'package:airqo/src/app/dashboard/widgets/air_quality_share_card.dart';
-import 'package:airqo/src/app/dashboard/widgets/clean_air_forum_camera_screen.dart';
-import 'package:airqo/src/app/dashboard/widgets/clean_air_forum_filter_card.dart';
+import 'package:airqo/src/app/dashboard/widgets/clean_air_forum_filter_tab.dart';
 import 'package:airqo/src/app/dashboard/widgets/clean_air_forum_sticker_frame.dart';
+import 'package:airqo/src/app/dashboard/widgets/share_sheet_widgets.dart';
 import 'package:airqo/src/app/learn/theme/learn_design_tokens.dart';
 import 'package:airqo/src/app/learn/widgets/learn_sheet_button_styles.dart';
 import 'package:airqo/src/app/shared/services/air_quality_share_service.dart';
 import 'package:airqo/src/app/shared/services/analytics_service.dart';
 import 'package:airqo/src/app/shared/services/clean_air_forum_submission_service.dart';
-import 'package:airqo/src/app/shared/widgets/custom_switch.dart';
 import 'package:airqo/src/meta/utils/colors.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:pasteboard/pasteboard.dart';
 import 'package:share_plus/share_plus.dart';
 
@@ -58,8 +53,6 @@ enum _ShareTab {
   const _ShareTab({required this.label, required this.analyticsName});
 }
 
-enum _SelfieSource { liveCamera, gallery }
-
 class AirQualityShareSheet extends StatefulWidget {
   final Measurement measurement;
 
@@ -91,17 +84,14 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
   _ShareTab _tab = _ShareTab.filter;
 
   final GlobalKey _cardKey = GlobalKey();
-  final GlobalKey _filterKey = GlobalKey();
   final GlobalKey _stickerKey = GlobalKey();
 
-  final ImagePicker _picker = ImagePicker();
-
+  // Lifted from the filter tab: its subtree is unmounted on tab switch, so
+  // the picked selfie and consent choice live here to survive switching.
   File? _selfieFile;
   bool _consentToDisplay = false;
 
-  bool _isPickingSelfie = false;
   bool _isSharingCard = false;
-  bool _isSharingFilter = false;
   bool _isCopyingSticker = false;
   bool _stickerCopied = false;
 
@@ -138,22 +128,6 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
     super.dispose();
   }
 
-  Future<Uint8List?> _captureBoundary(GlobalKey key) async {
-    final pixelRatio =
-        MediaQuery.of(context).devicePixelRatio.clamp(2.0, 3.0).toDouble();
-
-    await Future<void>.delayed(const Duration(milliseconds: 16));
-
-    final boundary =
-        key.currentContext?.findRenderObject() as RenderRepaintBoundary?;
-    if (boundary == null) return null;
-
-    final image = await boundary.toImage(pixelRatio: pixelRatio);
-    final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-
-    return byteData?.buffer.asUint8List();
-  }
-
   /// Shows an inline banner rather than a `SnackBar` — this sheet is a
   /// modal route above the app's own Scaffold, so a SnackBar raised via
   /// `ScaffoldMessenger` renders underneath it and is never actually seen.
@@ -171,7 +145,7 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
     setState(() => _isSharingCard = true);
 
     try {
-      final imageBytes = await _captureBoundary(_cardKey);
+      final imageBytes = await captureShareBoundary(context, _cardKey);
       if (imageBytes == null) {
         _showMessage('Could not prepare the share card. Please try again.');
         return;
@@ -196,175 +170,6 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
     }
   }
 
-  Future<void> _pickSelfie(ImageSource source) async {
-    if (_isPickingSelfie) return;
-    setState(() => _isPickingSelfie = true);
-
-    try {
-      final picked = await _picker.pickImage(
-        source: source,
-        maxWidth: 1440,
-        imageQuality: 90,
-        preferredCameraDevice: CameraDevice.front,
-      );
-      if (picked == null) return;
-      if (!mounted) return;
-
-      setState(() => _selfieFile = File(picked.path));
-    } catch (_) {
-      _showMessage('Could not access the camera/gallery. Please try again.');
-    } finally {
-      if (mounted) setState(() => _isPickingSelfie = false);
-    }
-  }
-
-  Future<void> _showSelfieSourceSheet() async {
-    final choice = await showModalBottomSheet<_SelfieSource>(
-      context: context,
-      backgroundColor: Colors.transparent,
-      builder: (sheetContext) {
-        // Same flush-to-bottom treatment as the main sheet — fold the
-        // safe-area inset into the content's own padding rather than
-        // leaving a transparent gap below the sheet.
-        final bottomSafeArea = MediaQuery.paddingOf(sheetContext).bottom;
-
-        return DecoratedBox(
-          decoration: BoxDecoration(
-            color: AppSurfaceColors.sheet(sheetContext),
-            borderRadius: const BorderRadius.vertical(
-              top: Radius.circular(LearnDesignTokens.sheetTopRadius),
-            ),
-          ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              LearnDesignTokens.dragHandle(sheetContext),
-              Padding(
-                padding: EdgeInsets.only(bottom: 8 + bottomSafeArea),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    _SelfieSourceTile(
-                      iconAsset: 'assets/icons/camera.svg',
-                      title: 'Take photo',
-                      subtitle: 'See the branding while you frame your shot',
-                      onTap: () => Navigator.of(sheetContext)
-                          .pop(_SelfieSource.liveCamera),
-                    ),
-                    _SelfieSourceTile(
-                      iconAsset: 'assets/icons/gallery.svg',
-                      title: 'Choose from gallery',
-                      onTap: () =>
-                          Navigator.of(sheetContext).pop(_SelfieSource.gallery),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-
-    switch (choice) {
-      case _SelfieSource.liveCamera:
-        AnalyticsService().trackCafSelfieSourceSelected(source: 'live_camera');
-        await _openLiveCamera();
-        break;
-      case _SelfieSource.gallery:
-        AnalyticsService().trackCafSelfieSourceSelected(source: 'gallery');
-        await _pickSelfie(ImageSource.gallery);
-        break;
-      case null:
-        break;
-    }
-  }
-
-  Future<void> _openLiveCamera() async {
-    if (_isPickingSelfie) return;
-    setState(() => _isPickingSelfie = true);
-
-    try {
-      final file = await Navigator.of(context).push<File>(
-        MaterialPageRoute(
-          settings: const RouteSettings(name: 'clean_air_forum_camera'),
-          fullscreenDialog: true,
-          builder: (_) => CleanAirForumCameraScreen(
-            measurement: widget.measurement,
-            fallbackLocationName: widget.fallbackLocationName,
-          ),
-        ),
-      );
-      if (file != null && mounted) {
-        setState(() => _selfieFile = file);
-        AnalyticsService().trackCafSelfieCaptured();
-      }
-    } catch (_) {
-      _showMessage('Could not open the live camera. Please try again.');
-    } finally {
-      if (mounted) setState(() => _isPickingSelfie = false);
-    }
-  }
-
-  Future<void> _shareFilter() async {
-    if (_isSharingFilter || _selfieFile == null) return;
-    setState(() => _isSharingFilter = true);
-
-    try {
-      final imageBytes = await _captureBoundary(_filterKey);
-      if (imageBytes == null) {
-        _showMessage('Could not prepare the filter. Please try again.');
-        return;
-      }
-
-      final result = await AirQualityShareService.shareCleanAirForumFilter(
-        imageBytes,
-        widget.measurement,
-        fallbackLocationName: widget.fallbackLocationName,
-        sharePositionOrigin: widget.sharePositionOrigin,
-      );
-      if (result.status == ShareResultStatus.success) {
-        AnalyticsService().trackCafFilterShared();
-        AnalyticsService().trackShareCompleted(
-          format: _ShareTab.filter.analyticsName,
-          method: 'share_sheet',
-        );
-      }
-
-      if (_consentToDisplay) {
-        unawaited(_submitToConferenceWall(imageBytes));
-      }
-    } catch (_) {
-      _showMessage('Could not share the filter. Please try again.');
-    } finally {
-      if (mounted) setState(() => _isSharingFilter = false);
-    }
-  }
-
-  CleanAirForumSubmissionService get _submissionService =>
-      widget.submissionService ?? CleanAirForumSubmissionService.instance;
-
-  Future<void> _submitToConferenceWall(Uint8List imageBytes) async {
-    try {
-      await _submissionService.submitSelfie(
-        imageBytes: imageBytes,
-        measurement: widget.measurement,
-        fallbackLocationName: widget.fallbackLocationName,
-      );
-      AnalyticsService().trackCafWallSubmissionSent();
-      _showMessage('Sent to the Clean Air Forum screen!');
-    } catch (e) {
-      // Report only the exception type — raw messages can carry API
-      // response details that don't belong in analytics.
-      AnalyticsService()
-          .trackCafWallSubmissionFailed(error: e.runtimeType.toString());
-      _showMessage(
-        'Your photo shared fine, but sending it to the conference screen '
-        "didn't work. Please try again.",
-      );
-    }
-  }
-
   /// Copies the sticker straight to the system clipboard so it can be
   /// pasted directly into an Instagram Story (or anywhere else) without
   /// going through the share sheet or saving a file first.
@@ -373,7 +178,7 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
     setState(() => _isCopyingSticker = true);
 
     try {
-      final imageBytes = await _captureBoundary(_stickerKey);
+      final imageBytes = await captureShareBoundary(context, _stickerKey);
       if (imageBytes == null) {
         _showMessage('Could not prepare the sticker. Please try again.');
         return;
@@ -459,7 +264,7 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
                       _buildTabSelector(),
                       if (_inlineMessage != null) ...[
                         const SizedBox(height: 12),
-                        _InlineMessageBanner(message: _inlineMessage!),
+                        InlineMessageBanner(message: _inlineMessage!),
                       ],
                       const SizedBox(height: 20),
                       _buildTabContent(),
@@ -491,7 +296,7 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
         for (final tab in _ShareTab.values) ...[
           if (tab != _ShareTab.values.first) const SizedBox(width: 8),
           Expanded(
-            child: _TabChip(
+            child: ShareTabChip(
               label: tab.label,
               selected: _tab == tab,
               onTap: () => _selectTab(tab),
@@ -505,7 +310,19 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
   Widget _buildTabContent() {
     switch (_tab) {
       case _ShareTab.filter:
-        return _buildFilterTab();
+        return CleanAirForumFilterTab(
+          measurement: widget.measurement,
+          fallbackLocationName: widget.fallbackLocationName,
+          sharePositionOrigin: widget.sharePositionOrigin,
+          analyticsFormat: _ShareTab.filter.analyticsName,
+          selfieFile: _selfieFile,
+          onSelfieChanged: (file) => setState(() => _selfieFile = file),
+          consentToDisplay: _consentToDisplay,
+          onConsentChanged: (value) =>
+              setState(() => _consentToDisplay = value),
+          onMessage: _showMessage,
+          submissionService: widget.submissionService,
+        );
       case _ShareTab.card:
         return _buildCardTab();
       case _ShareTab.sticker:
@@ -525,120 +342,10 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
           ),
         ),
         const SizedBox(height: 16),
-        _ShareButton(
+        ShareActionButton(
           label: _isSharingCard ? 'Preparing card...' : 'Share card',
           loading: _isSharingCard,
           onPressed: _isSharingCard ? null : _shareCard,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildFilterTab() {
-    final hasSelfie = _selfieFile != null;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        // The template is always visible — even before a selfie is picked —
-        // so it stays previewable while switching between share tabs. Full
-        // width (no artificial max-width cap) so it matches the Card tab's
-        // size for a consistent look across all three share formats.
-        RepaintBoundary(
-          key: _filterKey,
-          child: CleanAirForumFilterCard(
-            selfieFile: _selfieFile,
-            measurement: widget.measurement,
-            fallbackLocationName: widget.fallbackLocationName,
-          ),
-        ),
-        const SizedBox(height: 16),
-        if (hasSelfie)
-          OutlinedButton.icon(
-            onPressed: _isPickingSelfie ? null : _showSelfieSourceSheet,
-            style: learnExposureSecondaryButtonStyle(context),
-            icon: SvgPicture.asset(
-              'assets/icons/camera.svg',
-              width: 18,
-              height: 18,
-              colorFilter: ColorFilter.mode(
-                LearnDesignTokens.headline(context),
-                BlendMode.srcIn,
-              ),
-            ),
-            label: const Text('Change photo'),
-          )
-        else
-          ElevatedButton.icon(
-            onPressed: _isPickingSelfie ? null : _showSelfieSourceSheet,
-            style: learnExposurePrimaryButtonStyle(enabled: !_isPickingSelfie),
-            icon: _isPickingSelfie
-                ? const SizedBox(
-                    width: 16,
-                    height: 16,
-                    child: CircularProgressIndicator(
-                      strokeWidth: 2,
-                      color: Colors.white,
-                    ),
-                  )
-                : SvgPicture.asset(
-                    'assets/icons/camera.svg',
-                    width: 18,
-                    height: 18,
-                    colorFilter: ColorFilter.mode(
-                      _isPickingSelfie
-                          ? AppColors.boldHeadlineColor3
-                          : Colors.white,
-                      BlendMode.srcIn,
-                    ),
-                  ),
-            label: const Text('Take selfie'),
-          ),
-        if (hasSelfie) ...[
-          const SizedBox(height: 12),
-          _buildConsentSection(),
-          const SizedBox(height: 16),
-          _ShareButton(
-            label: _isSharingFilter ? 'Preparing filter...' : 'Share filter',
-            loading: _isSharingFilter,
-            onPressed: _isSharingFilter ? null : _shareFilter,
-          ),
-        ],
-      ],
-    );
-  }
-
-  Widget _buildConsentSection() {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                'Display my photo on the conference screen',
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.w600,
-                  color: LearnDesignTokens.headline(context),
-                ),
-              ),
-              const SizedBox(height: 2),
-              Text(
-                'Shown live on the Clean Air Forum wall display.',
-                style: LearnDesignTokens.completionCaption(context),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(width: 12),
-        CustomSwitch(
-          value: _consentToDisplay,
-          onChanged: (value) {
-            setState(() => _consentToDisplay = value);
-            if (value) AnalyticsService().trackCafWallConsentGiven();
-          },
         ),
       ],
     );
@@ -648,7 +355,7 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
       children: [
-        _CheckerboardBackground(
+        CheckerboardBackground(
           borderRadius: 26,
           child: RepaintBoundary(
             key: _stickerKey,
@@ -710,206 +417,4 @@ class _AirQualityShareSheetState extends State<AirQualityShareSheet> {
       ],
     );
   }
-}
-
-class _InlineMessageBanner extends StatelessWidget {
-  final String message;
-
-  const _InlineMessageBanner({required this.message});
-
-  @override
-  Widget build(BuildContext context) {
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: LearnDesignTokens.nestedSurface(context),
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 10),
-        child: Text(
-          message,
-          style: TextStyle(
-            fontSize: 13,
-            fontWeight: FontWeight.w600,
-            color: LearnDesignTokens.headline(context),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _SelfieSourceTile extends StatelessWidget {
-  final String iconAsset;
-  final String title;
-  final String? subtitle;
-  final VoidCallback onTap;
-
-  const _SelfieSourceTile({
-    required this.iconAsset,
-    required this.title,
-    this.subtitle,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      onTap: onTap,
-      leading: CircleAvatar(
-        backgroundColor: LearnDesignTokens.iconBg(context),
-        child: SvgPicture.asset(
-          iconAsset,
-          width: 20,
-          height: 20,
-          colorFilter: ColorFilter.mode(
-            LearnDesignTokens.headline(context),
-            BlendMode.srcIn,
-          ),
-        ),
-      ),
-      title: Text(
-        title,
-        style: TextStyle(
-          fontWeight: FontWeight.w600,
-          color: LearnDesignTokens.headline(context),
-        ),
-      ),
-      subtitle: subtitle == null
-          ? null
-          : Text(subtitle!,
-              style: LearnDesignTokens.completionCaption(context)),
-    );
-  }
-}
-
-/// Matches the pill-shaped view/country selector used on the dashboard, map,
-/// and learn tab (see `ViewSelector._buildViewButton`).
-class _TabChip extends StatelessWidget {
-  final String label;
-  final bool selected;
-  final VoidCallback onTap;
-
-  const _TabChip({
-    required this.label,
-    required this.selected,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        padding: const EdgeInsets.symmetric(vertical: 12),
-        alignment: Alignment.center,
-        decoration: BoxDecoration(
-          color: selected
-              ? AppColors.primaryColor
-              : (isDark
-                  ? AppColors.darkHighlight
-                  : AppColors.dividerColorlight),
-          borderRadius: BorderRadius.circular(30),
-        ),
-        child: Text(
-          label,
-          textAlign: TextAlign.center,
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-          style: TextStyle(
-            fontWeight: FontWeight.w500,
-            color: selected
-                ? Colors.white
-                : (isDark ? Colors.white : Colors.black87),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ShareButton extends StatelessWidget {
-  final String label;
-  final bool loading;
-  final VoidCallback? onPressed;
-
-  const _ShareButton({
-    required this.label,
-    required this.loading,
-    required this.onPressed,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton.icon(
-      onPressed: onPressed,
-      style: learnExposurePrimaryButtonStyle(enabled: onPressed != null),
-      icon: loading
-          ? const SizedBox(
-              width: 16,
-              height: 16,
-              child: CircularProgressIndicator(
-                strokeWidth: 2,
-                color: Colors.white,
-              ),
-            )
-          : SvgPicture.asset(
-              'assets/icons/share-icon.svg',
-              width: 18,
-              height: 18,
-              colorFilter:
-                  const ColorFilter.mode(Colors.white, BlendMode.srcIn),
-            ),
-      label: Text(label),
-    );
-  }
-}
-
-/// Simple checkerboard backdrop shown only in-app (not part of the captured
-/// image) so users can see the sticker preview really is transparent.
-class _CheckerboardBackground extends StatelessWidget {
-  final Widget child;
-  final double borderRadius;
-
-  const _CheckerboardBackground({required this.child, this.borderRadius = 0});
-
-  @override
-  Widget build(BuildContext context) {
-    return ClipRRect(
-      borderRadius: BorderRadius.circular(borderRadius),
-      child: CustomPaint(
-        painter: _CheckerboardPainter(),
-        child: Padding(
-          padding: const EdgeInsets.all(16),
-          child: child,
-        ),
-      ),
-    );
-  }
-}
-
-class _CheckerboardPainter extends CustomPainter {
-  static const double _tile = 10;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final light = Paint()..color = const Color(0xFFEDEDED);
-    final dark = Paint()..color = const Color(0xFFD8D8D8);
-
-    canvas.drawRect(Offset.zero & size, light);
-
-    for (double y = 0; y < size.height; y += _tile) {
-      for (double x = 0; x < size.width; x += _tile) {
-        final isDark = ((x / _tile).floor() + (y / _tile).floor()) % 2 == 0;
-        if (isDark) {
-          canvas.drawRect(Rect.fromLTWH(x, y, _tile, _tile), dark);
-        }
-      }
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
 }
